feat(investment): add currency display conversion for transactions
- Add display fields to InvestmentTransaction proto (displayPrice, displayCost, displayFees)
- Implement enrichTransactionProto to convert prices to user's preferred currency
- Update InvestmentDetailModal to show converted transaction prices
- Ensure consistent multi-currency display across portfolio and transaction views

This allows users to view transaction history in their preferred currency
even when investments are denominated in different currencies.

+++ src/go-backend/domain/repository/investment_repository_impl.go
@@ -126,7 +126,7 @@ func (r *investmentRepository) ListByWalletID(ctx context.Context, walletID int3
 
 	// Apply type filter if specified (0 is unspecified, so we don't filter on it)
 	if typeFilter != v1.InvestmentType_INVESTMENT_TYPE_UNSPECIFIED && typeFilter != 0 {
-		query = query.Where("type = ?", typeFilter)
+		query = query.Where("type = ?", int32(typeFilter))
 	}
 
 	// Get total count
